Move some more stuff to the appropriate class and a small bit of refactor
