+++ letsencrypt-win-simple/DefaultRunner.cs
@@ -10,50 +10,48 @@ namespace LetsEncrypt.ACME.Simple
         public static void Auto(Target binding)
         {
             var auth = App.LetsEncryptService.Authorize(binding);
-            if (auth.Status == "valid")
+            if (auth.Status != "valid")
+                return;
+
+            var pfxFilename = App.LetsEncryptService.GetCertificate(binding);
+
+            if (App.Options.Test && !App.Options.Renew && !App.ConsoleService
+                .PromptYesNo($"\nDo you want to install the .pfx into the Certificate Store/ Central SSL Store?"))
+                    return;
+
+            if (!App.Options.CentralSsl)
             {
-                var pfxFilename = App.LetsEncryptService.GetCertificate(binding);
+                X509Store store;
+                X509Certificate2 certificate;
+                Log.Information("Installing Non-Central SSL Certificate in the certificate store");
+                App.CertificateService.InstallCertificate(binding, pfxFilename, out store, out certificate);
 
                 if (App.Options.Test && !App.Options.Renew)
-                {
-                    if (!App.ConsoleService.PromptYesNo($"\nDo you want to install the .pfx into the Certificate Store/ Central SSL Store?"))
+                    if (!App.ConsoleService.PromptYesNo($"\nDo you want to add/update the certificate to your server software?"))
                         return;
-                }
 
-                if (!App.Options.CentralSsl)
-                {
-                    X509Store store;
-                    X509Certificate2 certificate;
-                    Log.Information("Installing Non-Central SSL Certificate in the certificate store");
-                    App.CertificateService.InstallCertificate(binding, pfxFilename, out store, out certificate);
+                Log.Information("Installing Non-Central SSL Certificate in server software");
+                binding.Plugin.Install(binding, pfxFilename, store, certificate);
+                if (!App.Options.KeepExisting)
+                    App.CertificateService.UninstallCertificate(binding.Host, out store, certificate);
+            }
+            else if (!App.Options.Renew || !App.Options.KeepExisting)
+            {
+                //If it is using centralized SSL, renewing, and replacing existing it needs to replace the existing binding.
+                Log.Information("Updating new Central SSL Certificate");
+                binding.Plugin.Install(binding);
+            }
 
-                    if (App.Options.Test && !App.Options.Renew)
-                        if (!App.ConsoleService.PromptYesNo($"\nDo you want to add/update the certificate to your server software?"))
-                            return;
+            if (App.Options.Test && !App.Options.Renew)
+            {
+                if (!App.ConsoleService.PromptYesNo($"\nDo you want to automatically renew this certificate in {App.Options.RenewalPeriodDays} days? This will add a task scheduler task."))
+                    return;
+            }
 
-                    Log.Information("Installing Non-Central SSL Certificate in server software");
-                    binding.Plugin.Install(binding, pfxFilename, store, certificate);
-                    if (!App.Options.KeepExisting)
-                        App.CertificateService.UninstallCertificate(binding.Host, out store, certificate);
-                }
-                else if (!App.Options.Renew || !App.Options.KeepExisting)
-                {
-                    //If it is using centralized SSL, renewing, and replacing existing it needs to replace the existing binding.
-                    Log.Information("Updating new Central SSL Certificate");
-                    binding.Plugin.Install(binding);
-                }
-
-                if (App.Options.Test && !App.Options.Renew)
-                {
-                    if (!App.ConsoleService.PromptYesNo($"\nDo you want to automatically renew this certificate in {App.Options.RenewalPeriodDays} days? This will add a task scheduler task."))
-                        return;
-                }
-
-                if (!App.Options.Renew)
-                {
-                    Log.Information("Adding renewal for {binding}", binding);
-                    Scheduler.ScheduleRenewal(binding);
-                }
+            if (!App.Options.Renew)
+            {
+                Log.Information("Adding renewal for {binding}", binding);
+                Scheduler.ScheduleRenewal(binding);
             }
         }
     }
