+++ letsencrypt-win-simple/Program.cs
@@ -44,31 +44,7 @@ namespace LetsEncrypt.ACME.Simple
 
                             Console.WriteLine();
                             App.ConsoleService.PrintMenuForPlugins();
-
-                            if (string.IsNullOrEmpty(App.Options.ManualHost) && string.IsNullOrWhiteSpace(App.Options.Plugin))
-                            {
-                                Console.WriteLine(" A: Get certificates for all hosts");
-                                Console.WriteLine(" Q: Quit");
-                                Console.Write("Which host do you want to get a certificate for: ");
-                                var command = App.ConsoleService.ReadCommandFromConsole();
-                                switch (command)
-                                {
-                                    case "a":
-                                        App.CertificateService.GetCertificatesForAllHosts(targets);
-                                        break;
-                                    case "q":
-                                        return;
-                                    default:
-                                        Target.ProcessDefaultCommand(targets, command);
-                                        break;
-                                }
-                            }
-                            else if (!string.IsNullOrWhiteSpace(App.Options.Plugin))
-                            {
-                                // If there's a plugin in the options, only do ProcessDefaultCommand for the selected plugin
-                                // Plugins that can run automatically should allow for an empty string as menu response to work
-                                Target.ProcessDefaultCommand(targets, string.Empty);
-                            }
+                            App.ConsoleService.PrintMenu(targets);
                         }
                     }
