+++ letsencrypt-win-simple/Services/ConsoleService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using LetsEncrypt.ACME.Simple.Configuration;
 
 namespace LetsEncrypt.ACME.Simple.Services
@@ -23,7 +24,7 @@ namespace LetsEncrypt.ACME.Simple.Services
             }
             return false;
         }
-        
+
         public void PrintMenuForPlugins()
         {
             // Check for a plugin specified in the options
@@ -42,6 +43,34 @@ namespace LetsEncrypt.ACME.Simple.Services
                 {
                     plugin.PrintMenu();
                 }
+            }
+        }
+
+        public void PrintMenu(List<Target> targets)
+        {
+            if (string.IsNullOrEmpty(App.Options.ManualHost) && string.IsNullOrWhiteSpace(App.Options.Plugin))
+            {
+                Console.WriteLine(" A: Get certificates for all hosts");
+                Console.WriteLine(" Q: Quit");
+                Console.Write("Which host do you want to get a certificate for: ");
+                var command = App.ConsoleService.ReadCommandFromConsole();
+                switch (command)
+                {
+                    case "a":
+                        App.CertificateService.GetCertificatesForAllHosts(targets);
+                        break;
+                    case "q":
+                        return;
+                    default:
+                        Target.ProcessDefaultCommand(targets, command);
+                        break;
+                }
+            }
+            else if (!string.IsNullOrWhiteSpace(App.Options.Plugin))
+            {
+                // If there's a plugin in the options, only do ProcessDefaultCommand for the selected plugin
+                // Plugins that can run automatically should allow for an empty string as menu response to work
+                Target.ProcessDefaultCommand(targets, string.Empty);
             }
         }
     }
